refactor : 엔티티 연관관계 및 어노테이션 수정
