+++ src/main/java/kr/kernel360/anabada/E2e1AnabadaApplication.java
@@ -2,7 +2,9 @@ package kr.kernel360.anabada;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class E2e1AnabadaApplication {
