+++ src/main/java/kr/kernel360/anabada/domain/category/entity/Category.java
@@ -1,4 +1,4 @@
-package kr.kernel360.anabada.faq;
+package kr.kernel360.anabada.domain.category.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -14,15 +14,12 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "Faq")
-public class Faq {
+@Table(name = "category")
+public class Category {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@Column(nullable = false, name = "title", columnDefinition = "varchar(50)")
-	private String title;
-
-	@Column(name = "content", columnDefinition = "text")
-	private String text;
+	@Column(nullable = false, name = "name", columnDefinition = "varchar(50)")
+	private String name;
 }
