+++ src/main/java/kr/kernel360/anabada/domain/comment/entity/Comment.java
@@ -1,17 +1,16 @@
-package kr.kernel360.anabada.comment;
+package kr.kernel360.anabada.domain.comment.entity;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import kr.kernel360.anabada.member.Member;
-import kr.kernel360.anabada.trade.Trade;
+import kr.kernel360.anabada.domain.member.entity.Member;
+import kr.kernel360.anabada.domain.trade.entity.Trade;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -25,10 +24,11 @@ public class Comment {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-	@Column(nullable = false, name = "created_by")
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(nullable = false, name = "created_by", columnDefinition = "bigint(50)")
 	private Member member;
 
-	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(nullable = false, name = "trade_id", columnDefinition = "bigint(50)")
 	private Trade trade;
 }
