+++ src/main/java/kr/kernel360/anabada/domain/tradeoffer/entity/TradeOffer.java
@@ -1,18 +1,18 @@
-package kr.kernel360.anabada.tradeoffer;
+package kr.kernel360.anabada.domain.tradeoffer.entity;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import kr.kernel360.anabada.global.commons.entity.BaseEntity;
-import kr.kernel360.anabada.member.Member;
-import kr.kernel360.anabada.trade.Trade;
+import kr.kernel360.anabada.domain.member.entity.Member;
+import kr.kernel360.anabada.domain.trade.entity.Trade;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -26,14 +26,6 @@ public class TradeOffer extends BaseEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-	@Column(nullable = false)
-	private Member member;
-
-	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-	@Column(nullable = false)
-	private Trade trade;
-
 	@Column(nullable = false, name = "title", columnDefinition = "varchar(40)")
 	private String title;
 
@@ -45,4 +37,12 @@ public class TradeOffer extends BaseEntity {
 
 	@Column(nullable = false, name = "trade_offer_status", columnDefinition = "tinyint")
 	private Boolean tradeOfferStatus;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(nullable = false, name = "created_by", columnDefinition = "bigint(50)")
+	private Member member;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(nullable = false, name = "trade_id", columnDefinition = "bigint(50)")
+	private Trade trade;
 }
